fix: use node affinity objects instead of their keys when mapping
Previously, when a user specified node affinities,
for every provided node affinity
we would try to map every _key_ to an "affinity" block.
This produced invalid Terraform resources,
because the values of the "key", "name" and "operator" keys on their own
are not enough to create a valid "affinity" block.

This commit uses entire affinity objects
and maps them to appropriate Terraform representations, fixing the problem.