feat: allow specifying Castware API URL
This is necessary to be able to differentiate between the API URL used
by the Cast AI Terraform provider and the Castware components running in
the cluster.

The motivation for different URLs comes from accessing the Cast AI API
through GCP Private Service Connect. In that case the API URL accessed
privately will contain a `.psc.` suffix, but that is only accessible via
the private connections. Since Terraform will most probably not run with
access to the private connection, it needs to be able to use the "usual"
public URL, while setting up Castware components with the private URL.